Add emulator to project from macOS project - https://github.com/ryanggrey/chip-8-macos
Once this is working and optimised I'll look into the possibility of a
shared lib/framework.

Renders using SKScene and one SKShapdeNode that renders a path. The path
is a discontinuous representation of the chip8 pixels/bitmap. The rendering
seems a bit slow (in comparison to macOS) and I'm not sure if this is
emulator or rendering bound. Further work to try and optimise will follow.